wip moving, move up alive
