+++ backend/ws_game_control_up.go
@@ -6,107 +6,6 @@ import (
 	"time"
 )
 
-func ws_character_control_handler(client *Client, control string) {
-
-	switch control {
-	case string(WS_UP_ON):
-		ws_up_handler(string_number[client.NUMBER], control, true)
-	case string(WS_UP_OFF):
-		ws_up_off_handler(string_number[client.NUMBER], control)
-		log.Println("WS_UP_OFF", client.NUMBER)
-	case string(WS_DOWN_ON):
-		log.Println("WS_DOWN_ON", client.NUMBER)
-	case string(WS_DOWN_OFF):
-		log.Println("WS_DOWN_OFF", client.NUMBER)
-	case string(WS_LEFT_ON):
-		log.Println("WS_LEFT_ON", client.NUMBER)
-	case string(WS_LEFT_OFF):
-		log.Println("WS_LEFT_OFF", client.NUMBER)
-	case string(WS_RIGHT_ON):
-		log.Println("WS_RIGHT_ON", client.NUMBER)
-	case string(WS_RIGHT_OFF):
-		log.Println("WS_RIGHT_OFF", client.NUMBER)
-	case string(WS_BOMB_ON):
-		log.Println("WS_BOMB_ON", client.NUMBER)
-	case string(WS_BOMB_OFF):
-		log.Println("WS_BOMB_OFF", client.NUMBER)
-	default:
-		log.Println("Unknown control: ", control)
-	}
-}
-
-// move_cells_per_second, depends on turbo
-var move_cps = map[bool]int64{
-	true:  2, // turbo
-	false: 1,
-}
-
-func unpress_arrow(player *PLAYER, control string) {
-	switch control {
-	case string(WS_UP_OFF):
-		player.up_pressed = false
-	case string(WS_DOWN_OFF):
-		player.down_pressed = false
-	case string(WS_LEFT_OFF):
-		player.left_pressed = false
-	case string(WS_RIGHT_OFF):
-		player.right_pressed = false
-	default:
-		log.Println("Unknown control to unpress: ", control)
-	}
-}
-
-func unpress_all_arrows(player *PLAYER) {
-	player.up_pressed = false
-	player.down_pressed = false
-	player.left_pressed = false
-	player.right_pressed = false
-}
-
-// todo: check this . pointers or without.
-func press_arrow_unpress_other_arrows(player *PLAYER, control string) {
-	switch control {
-	case string(WS_UP_ON):
-		player.down_pressed = false
-		player.left_pressed = false
-		player.right_pressed = false
-		player.up_pressed = true
-	case string(WS_DOWN_ON):
-		player.up_pressed = false
-		player.left_pressed = false
-		player.right_pressed = false
-		player.down_pressed = true
-	case string(WS_LEFT_ON):
-		player.up_pressed = false
-		player.down_pressed = false
-		player.right_pressed = false
-		player.left_pressed = true
-	case string(WS_RIGHT_ON):
-		player.up_pressed = false
-		player.down_pressed = false
-		player.left_pressed = false
-		player.right_pressed = true
-	default:
-		log.Println("Unknown control to unpress arrows: ", control)
-	}
-}
-
-// executed as goroutine. listen for player press arrows
-func ws_arrows_loop_listener() {
-	for game_waiting_state == GAME_STARTED {
-		//sleep 0.1 sec
-		time.Sleep(100 * time.Millisecond)
-		game.Players.Range(func(key, value interface{}) bool {
-			number := key.(string)
-			player := value.(PLAYER)
-			if !player.Dead {
-				ws_up_handler(number, string(WS_UP_ON), false)
-			}
-			return true
-		})
-	}
-}
-
 func ws_up_off_handler(number string, control string) {
 	dprint("ws_up_off_handler. player", number)
 
@@ -137,9 +36,11 @@ func ws_up_handler(number string, control string, press bool) {
 	unix_ts := time.Now().UnixNano()
 
 	if !player.moving && press {
+		dprint("inside if !player.moving && press")
 		// Check if player can move up
 		target_position_cell := fmt.Sprintf("%d%d", player.X, player.Y-1)
 		if _, ok := game.free_cells.Load(target_position_cell); !ok {
+			dprint("target cell is not free", target_position_cell)
 			player.up_pressed = false
 			return
 		}
@@ -179,7 +80,8 @@ func ws_up_handler(number string, control string, press bool) {
 		game.Players.Store(number, player)
 		dprint("inside if time 80. before ws_send_move_up_command")
 		ws_send_move_up_command(&player)
-	} else if player.moving && player.can_change_present_cell &&
+	} else if player.moving &&
+		player.can_change_present_cell &&
 		player.one_cell_move_duration/10*5 < unix_ts-player.moving_start_time_stamp {
 		// Check if 50% of the timer is done, switch to the next cell
 		dprint("inside if time 50")
@@ -189,9 +91,11 @@ func ws_up_handler(number string, control string, press bool) {
 		game.Players.Store(number, player)
 	} else if player.moving &&
 		player.one_cell_move_duration < unix_ts-player.moving_start_time_stamp {
+		dprint("inside if time 100")
 		player.moving = false
 		player.can_change_present_cell = false
 		player.can_change_target_cell = false
+		game.Players.Store(number, player)
 		// dprint("50 time:", player.one_cell_move_duration/10*5 < unix_ts-player.moving_start_time_stamp, player.one_cell_move_duration/10*5/(unix_ts-player.moving_start_time_stamp))
 		// dprint("80 time:", player.one_cell_move_duration/10*8 < unix_ts-player.moving_start_time_stamp, player.one_cell_move_duration/10*8/(unix_ts-player.moving_start_time_stamp))
 	}
